requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
google-generativeai>=0.3.1
python-telegram-bot>=20.7
requests>=2.31.0
pydantic>=2.5.2
python-dotenv>=1.0.0
